requirements.txt
streamlit
pandas
geopandas
folium
streamlit-folium
plotly
numpy
requests
google-generativeai
matplotlib
shapely
